refactor import/export according mentor's advice
