+++ js/data/generate-ad.js
@@ -1,9 +1,9 @@
-import generateAuthor from './generate-author.js';
-import generateOffer from './generate-offer.js';
+import { generateAuthor } from './generate-author.js';
+import { generateOffer } from './generate-offer.js';
 
 const generateAd = () => ({
   author: generateAuthor(),
   offer: generateOffer(),
 });
 
-export default generateAd;
+export { generateAd };
