+++ js/utils/generate-random-int.js
@@ -10,4 +10,4 @@ const generateRandomInt = (min, max) => (
   Math.floor(Math.random() * (max - min + 1) + min)
 );
 
-export default generateRandomInt;
+export { generateRandomInt };
