+++ js/utils/index.js
@@ -1,0 +1,5 @@
+export * from './generate-array-of-strings.js';
+export * from './get-random-float.js';
+export * from './get-random-int.js';
+export * from './get-uniq-random-items-from-array.js';
+export * from './make-counter.js';
